fetch/requirements.txt
beautifulsoup4==4.5.1
Pillow==3.4.2
requests==2.11.1
